Move media-image into its own file
