+++ src/components/media-image.js
@@ -1,0 +1,212 @@
+import { createImageTexture } from "../utils/media-utils";
+import { createBasisTexture, createKTX2Texture } from "../utils/create-basis-texture";
+import { TextureCache } from "../utils/texture-cache";
+import { GIFTexture } from "../utils/gif-texture";
+import { errorTexture } from "../utils/error-texture";
+import { createPlaneBufferGeometry } from "../utils/three-utils";
+import { promisifyWorker } from "../utils/promisify-worker.js";
+import { scaleToAspectRatio } from "../utils/scale-to-aspect-ratio";
+import GIFWorker from "../workers/gifparsing.worker.js";
+
+const parseGIF = promisifyWorker(new GIFWorker());
+
+const textureCache = new TextureCache();
+const inflightTextures = new Map();
+
+const errorCacheItem = { texture: errorTexture, ratio: 1400 / 1200 };
+
+async function createGIFTexture(url) {
+  return new Promise((resolve, reject) => {
+    fetch(url, { mode: "cors" })
+      .then(r => r.arrayBuffer())
+      .then(rawImageData => parseGIF(rawImageData, [rawImageData]))
+      .then(result => {
+        const { frames, delayTimes, disposals } = result;
+        let loadCnt = 0;
+        for (let i = 0; i < frames.length; i++) {
+          const img = new Image();
+          img.onload = e => {
+            loadCnt++;
+            frames[i] = e.target;
+            if (loadCnt === frames.length) {
+              const texture = new GIFTexture(frames, delayTimes, disposals);
+              texture.image.src = url;
+              texture.encoding = THREE.sRGBEncoding;
+              texture.minFilter = THREE.LinearFilter;
+              resolve(texture);
+            }
+          };
+          img.src = frames[i];
+        }
+      })
+      .catch(reject);
+  });
+}
+
+AFRAME.registerComponent("media-image", {
+  schema: {
+    src: { type: "string" },
+    version: { type: "number" },
+    projection: { type: "string", default: "flat" },
+    contentType: { type: "string" },
+    batch: { default: false },
+    alphaMode: { type: "string", default: undefined },
+    alphaCutoff: { type: "number" }
+  },
+
+  remove() {
+    if (this.data.batch && this.mesh) {
+      this.el.sceneEl.systems["hubs-systems"].batchManagerSystem.removeObject(this.mesh);
+    }
+    if (this.currentSrcIsRetained) {
+      textureCache.release(this.data.src, this.data.version);
+      this.currentSrcIsRetained = false;
+    }
+  },
+
+  async update(oldData) {
+    let texture;
+    let ratio = 1;
+
+    const batchManagerSystem = this.el.sceneEl.systems["hubs-systems"].batchManagerSystem;
+
+    try {
+      const { src, version, contentType } = this.data;
+      if (!src) return;
+
+      this.el.emit("image-loading");
+
+      if (this.mesh && this.mesh.material.map && (src !== oldData.src || version !== oldData.version)) {
+        this.mesh.material.map = null;
+        this.mesh.material.needsUpdate = true;
+        if (this.mesh.material.map !== errorTexture) {
+          textureCache.release(oldData.src, oldData.version);
+          this.currentSrcIsRetained = false;
+        }
+      }
+
+      let cacheItem;
+      if (textureCache.has(src, version)) {
+        if (this.currentSrcIsRetained) {
+          cacheItem = textureCache.get(src, version);
+        } else {
+          cacheItem = textureCache.retain(src, version);
+        }
+      } else {
+        const inflightKey = textureCache.key(src, version);
+
+        if (src === "error") {
+          cacheItem = errorCacheItem;
+        } else if (inflightTextures.has(inflightKey)) {
+          await inflightTextures.get(inflightKey);
+          cacheItem = textureCache.retain(src, version);
+        } else {
+          let promise;
+          if (contentType.includes("image/gif")) {
+            promise = createGIFTexture(src);
+          } else if (contentType.includes("image/basis")) {
+            promise = createBasisTexture(src);
+          } else if (contentType.includes("image/ktx2")) {
+            promise = createKTX2Texture(src);
+          } else if (contentType.startsWith("image/")) {
+            promise = createImageTexture(src);
+          } else {
+            throw new Error(`Unknown image content type: ${contentType}`);
+          }
+          inflightTextures.set(inflightKey, promise);
+          texture = await promise;
+          inflightTextures.delete(inflightKey);
+          cacheItem = textureCache.set(src, version, texture);
+        }
+
+        // No way to cancel promises, so if src has changed or this entity was removed while we were creating the texture just throw it away.
+        if (this.data.src !== src || this.data.version !== version || !this.el.parentNode) {
+          textureCache.release(src, version);
+          return;
+        }
+      }
+
+      texture = cacheItem.texture;
+      ratio = cacheItem.ratio;
+
+      this.currentSrcIsRetained = true;
+    } catch (e) {
+      console.error("Error loading image", this.data.src, e);
+      texture = errorTexture;
+      this.currentSrcIsRetained = false;
+    }
+
+    const projection = this.data.projection;
+
+    if (this.mesh && this.data.batch) {
+      // This is a no-op if the mesh was just created.
+      // Otherwise we want to ensure the texture gets updated.
+      batchManagerSystem.removeObject(this.mesh);
+    }
+
+    if (!this.mesh || projection !== oldData.projection) {
+      const material = new THREE.MeshBasicMaterial();
+
+      let geometry;
+
+      if (projection === "360-equirectangular") {
+        geometry = new THREE.SphereBufferGeometry(1, 64, 32);
+        // invert the geometry on the x-axis so that all of the faces point inward
+        geometry.scale(-1, 1, 1);
+
+        // Flip uvs on the geometry
+        if (!texture.flipY) {
+          const uvs = geometry.attributes.uv.array;
+
+          for (let i = 1; i < uvs.length; i += 2) {
+            uvs[i] = 1 - uvs[i];
+          }
+        }
+      } else {
+        geometry = createPlaneBufferGeometry(1, 1, 1, 1, texture.flipY);
+        material.side = THREE.DoubleSide;
+      }
+
+      this.mesh = new THREE.Mesh(geometry, material);
+      this.el.setObject3D("mesh", this.mesh);
+    }
+
+    if (texture == errorTexture) {
+      this.mesh.material.transparent = true;
+    } else {
+      // if transparency setting isnt explicitly defined, default to on for all non batched things, gifs, and basis textures with alpha
+      switch (this.data.alphaMode) {
+        case "opaque":
+          this.mesh.material.transparent = false;
+          break;
+        case "blend":
+          this.mesh.material.transparent = true;
+          this.mesh.material.alphaTest = 0;
+          break;
+        case "mask":
+          this.mesh.material.transparent = false;
+          this.mesh.material.alphaTest = this.data.alphaCutoff;
+          break;
+        default:
+          this.mesh.material.transparent =
+            !this.data.batch ||
+            this.data.contentType.includes("image/gif") ||
+            !!(texture.image && texture.image.hasAlpha);
+          this.mesh.material.alphaTest = 0;
+      }
+    }
+
+    this.mesh.material.map = texture;
+    this.mesh.material.needsUpdate = true;
+
+    if (projection === "flat") {
+      scaleToAspectRatio(this.el, ratio);
+    }
+
+    if (texture !== errorTexture && this.data.batch && !texture.isCompressedTexture) {
+      batchManagerSystem.addObject(this.mesh);
+    }
+
+    this.el.emit("image-loaded", { src: this.data.src, projection: projection });
+  }
+});
