+++ src/components/media-views.js
@@ -1,16 +1,14 @@
 /* global performance THREE AFRAME NAF MediaStream setTimeout */
 import configs from "../utils/configs";
-import GIFWorker from "../workers/gifparsing.worker.js";
 import audioIcon from "../assets/images/audio.png";
 import { paths } from "../systems/userinput/paths";
 import HLS from "hls.js";
 import { MediaPlayer } from "dashjs";
-import { addAndArrangeMedia, createImageTexture, createVideoOrAudioEl } from "../utils/media-utils";
+import { addAndArrangeMedia, createVideoOrAudioEl } from "../utils/media-utils";
 import { disposeTexture } from "../utils/material-utils";
 import { proxiedUrlFor } from "../utils/media-url-utils";
 import { buildAbsoluteURL } from "url-toolkit";
 import { SOUND_CAMERA_TOOL_TOOK_SNAPSHOT } from "../systems/sound-effects-system";
-import { promisifyWorker } from "../utils/promisify-worker.js";
 import { applyPersistentSync } from "../utils/permissions-utils";
 import { refreshMediaMirror, getCurrentMirroredMedia } from "../utils/mirror-utils";
 import { detect } from "detect-browser";
@@ -26,7 +24,6 @@ import qsTruthy from "../utils/qs_truthy";
 
 const ONCE_TRUE = { once: true };
 const TYPE_IMG_PNG = { type: "image/png" };
-const parseGIF = promisifyWorker(new GIFWorker());
 
 const isIOS = AFRAME.utils.device.isIOS();
 const audioIconTexture = new HubsTextureLoader().load(audioIcon);
@@ -40,181 +37,6 @@ for (let i = 0; i <= 20; i++) {
   s += "]";
   VOLUME_LABELS[i] = s;
 }
-
-import { KTX2Loader } from "three/examples/jsm/loaders/KTX2Loader";
-import { rewriteBasisTranscoderUrls } from "../utils/media-url-utils";
-const loadingManager = new THREE.LoadingManager();
-loadingManager.setURLModifier(rewriteBasisTranscoderUrls);
-
-let ktxLoader;
-
-export function createBasisTexture(url) {
-  if (!ktxLoader) {
-    ktxLoader = new KTX2Loader(loadingManager).detectSupport(AFRAME.scenes[0].renderer);
-  }
-  return new Promise((resolve, reject) => {
-    ktxLoader.basisLoader.load(
-      url,
-      function(texture) {
-        texture.encoding = THREE.sRGBEncoding;
-        texture.onUpdate = function() {
-          // Delete texture data once it has been uploaded to the GPU
-          texture.mipmaps.length = 0;
-        };
-        // texture.anisotropy = 4;
-        resolve(texture);
-      },
-      undefined,
-      function(error) {
-        console.error(error);
-        reject(new Error(`'${url}' could not be fetched (Error: ${error}`));
-      }
-    );
-  });
-}
-
-export function createKTX2Texture(url) {
-  if (!ktxLoader) {
-    ktxLoader = new KTX2Loader(loadingManager).detectSupport(AFRAME.scenes[0].renderer);
-  }
-  return new Promise((resolve, reject) => {
-    ktxLoader.load(
-      url,
-      function(texture) {
-        texture.encoding = THREE.sRGBEncoding;
-        texture.onUpdate = function() {
-          // Delete texture data once it has been uploaded to the GPU
-          texture.mipmaps.length = 0;
-        };
-        texture.anisotropy = 4;
-        resolve(texture);
-      },
-      undefined,
-      function(error) {
-        console.error(error);
-        reject(new Error(`'${url}' could not be fetched (Error: ${error}`));
-      }
-    );
-  });
-}
-
-class GIFTexture extends THREE.Texture {
-  constructor(frames, delays, disposals) {
-    super(document.createElement("canvas"));
-    this.image.width = frames[0].width;
-    this.image.height = frames[0].height;
-
-    this._ctx = this.image.getContext("2d");
-
-    this.generateMipmaps = false;
-    this.isVideoTexture = true;
-    this.minFilter = THREE.NearestFilter;
-
-    this.frames = frames;
-    this.delays = delays;
-    this.disposals = disposals;
-
-    this.frame = 0;
-    this.frameStartTime = Date.now();
-  }
-
-  update() {
-    if (!this.frames || !this.delays || !this.disposals) return;
-    const now = Date.now();
-    if (now - this.frameStartTime > this.delays[this.frame]) {
-      if (this.disposals[this.frame] === 2) {
-        this._ctx.clearRect(0, 0, this.image.width, this.image.width);
-      }
-      this.frame = (this.frame + 1) % this.frames.length;
-      this.frameStartTime = now;
-      this._ctx.drawImage(this.frames[this.frame], 0, 0, this.image.width, this.image.height);
-      this.needsUpdate = true;
-    }
-  }
-}
-
-async function createGIFTexture(url) {
-  return new Promise((resolve, reject) => {
-    fetch(url, { mode: "cors" })
-      .then(r => r.arrayBuffer())
-      .then(rawImageData => parseGIF(rawImageData, [rawImageData]))
-      .then(result => {
-        const { frames, delayTimes, disposals } = result;
-        let loadCnt = 0;
-        for (let i = 0; i < frames.length; i++) {
-          const img = new Image();
-          img.onload = e => {
-            loadCnt++;
-            frames[i] = e.target;
-            if (loadCnt === frames.length) {
-              const texture = new GIFTexture(frames, delayTimes, disposals);
-              texture.image.src = url;
-              texture.encoding = THREE.sRGBEncoding;
-              texture.minFilter = THREE.LinearFilter;
-              resolve(texture);
-            }
-          };
-          img.src = frames[i];
-        }
-      })
-      .catch(reject);
-  });
-}
-
-class TextureCache {
-  cache = new Map();
-
-  key(src, version) {
-    return `${src}_${version}`;
-  }
-
-  set(src, version, texture) {
-    const image = texture.image;
-    this.cache.set(this.key(src, version), {
-      texture,
-      ratio: (image.videoHeight || image.height) / (image.videoWidth || image.width),
-      count: 0
-    });
-    return this.retain(src, version);
-  }
-
-  has(src, version) {
-    return this.cache.has(this.key(src, version));
-  }
-
-  get(src, version) {
-    return this.cache.get(this.key(src, version));
-  }
-
-  retain(src, version) {
-    const cacheItem = this.cache.get(this.key(src, version));
-    cacheItem.count++;
-    // console.log("retain", src, cacheItem.count);
-    return cacheItem;
-  }
-
-  release(src, version) {
-    const cacheItem = this.cache.get(this.key(src, version));
-
-    if (!cacheItem) {
-      console.error(`Releasing uncached texture src ${src}`);
-      return;
-    }
-
-    cacheItem.count--;
-    // console.log("release", src, cacheItem.count);
-    if (cacheItem.count <= 0) {
-      // Unload the video element to prevent it from continuing to play in the background
-      disposeTexture(cacheItem.texture);
-      this.cache.delete(this.key(src, version));
-    }
-  }
-}
-
-const textureCache = new TextureCache();
-const inflightTextures = new Map();
-
-const errorCacheItem = { texture: errorTexture, ratio: 1400 / 1200 };
 
 function timeFmt(t) {
   let s = Math.floor(t),
@@ -997,173 +819,5 @@ AFRAME.registerComponent("media-video", {
     }
 
     window.APP.store.removeEventListener("statechanged", this.onPreferenceChanged);
-  }
-});
-
-AFRAME.registerComponent("media-image", {
-  schema: {
-    src: { type: "string" },
-    version: { type: "number" },
-    projection: { type: "string", default: "flat" },
-    contentType: { type: "string" },
-    batch: { default: false },
-    alphaMode: { type: "string", default: undefined },
-    alphaCutoff: { type: "number" }
-  },
-
-  remove() {
-    if (this.data.batch && this.mesh) {
-      this.el.sceneEl.systems["hubs-systems"].batchManagerSystem.removeObject(this.mesh);
-    }
-    if (this.currentSrcIsRetained) {
-      textureCache.release(this.data.src, this.data.version);
-      this.currentSrcIsRetained = false;
-    }
-  },
-
-  async update(oldData) {
-    let texture;
-    let ratio = 1;
-
-    const batchManagerSystem = this.el.sceneEl.systems["hubs-systems"].batchManagerSystem;
-
-    try {
-      const { src, version, contentType } = this.data;
-      if (!src) return;
-
-      this.el.emit("image-loading");
-
-      if (this.mesh && this.mesh.material.map && (src !== oldData.src || version !== oldData.version)) {
-        this.mesh.material.map = null;
-        this.mesh.material.needsUpdate = true;
-        if (this.mesh.material.map !== errorTexture) {
-          textureCache.release(oldData.src, oldData.version);
-          this.currentSrcIsRetained = false;
-        }
-      }
-
-      let cacheItem;
-      if (textureCache.has(src, version)) {
-        if (this.currentSrcIsRetained) {
-          cacheItem = textureCache.get(src, version);
-        } else {
-          cacheItem = textureCache.retain(src, version);
-        }
-      } else {
-        const inflightKey = textureCache.key(src, version);
-
-        if (src === "error") {
-          cacheItem = errorCacheItem;
-        } else if (inflightTextures.has(inflightKey)) {
-          await inflightTextures.get(inflightKey);
-          cacheItem = textureCache.retain(src, version);
-        } else {
-          let promise;
-          if (contentType.includes("image/gif")) {
-            promise = createGIFTexture(src);
-          } else if (contentType.includes("image/basis")) {
-            promise = createBasisTexture(src);
-          } else if (contentType.includes("image/ktx2")) {
-            promise = createKTX2Texture(src);
-          } else if (contentType.startsWith("image/")) {
-            promise = createImageTexture(src);
-          } else {
-            throw new Error(`Unknown image content type: ${contentType}`);
-          }
-          inflightTextures.set(inflightKey, promise);
-          texture = await promise;
-          inflightTextures.delete(inflightKey);
-          cacheItem = textureCache.set(src, version, texture);
-        }
-
-        // No way to cancel promises, so if src has changed or this entity was removed while we were creating the texture just throw it away.
-        if (this.data.src !== src || this.data.version !== version || !this.el.parentNode) {
-          textureCache.release(src, version);
-          return;
-        }
-      }
-
-      texture = cacheItem.texture;
-      ratio = cacheItem.ratio;
-
-      this.currentSrcIsRetained = true;
-    } catch (e) {
-      console.error("Error loading image", this.data.src, e);
-      texture = errorTexture;
-      this.currentSrcIsRetained = false;
-    }
-
-    const projection = this.data.projection;
-
-    if (this.mesh && this.data.batch) {
-      // This is a no-op if the mesh was just created.
-      // Otherwise we want to ensure the texture gets updated.
-      batchManagerSystem.removeObject(this.mesh);
-    }
-
-    if (!this.mesh || projection !== oldData.projection) {
-      const material = new THREE.MeshBasicMaterial();
-
-      let geometry;
-
-      if (projection === "360-equirectangular") {
-        geometry = new THREE.SphereBufferGeometry(1, 64, 32);
-        // invert the geometry on the x-axis so that all of the faces point inward
-        geometry.scale(-1, 1, 1);
-
-        // Flip uvs on the geometry
-        if (!texture.flipY) {
-          const uvs = geometry.attributes.uv.array;
-
-          for (let i = 1; i < uvs.length; i += 2) {
-            uvs[i] = 1 - uvs[i];
-          }
-        }
-      } else {
-        geometry = createPlaneBufferGeometry(1, 1, 1, 1, texture.flipY);
-        material.side = THREE.DoubleSide;
-      }
-
-      this.mesh = new THREE.Mesh(geometry, material);
-      this.el.setObject3D("mesh", this.mesh);
-    }
-
-    if (texture == errorTexture) {
-      this.mesh.material.transparent = true;
-    } else {
-      // if transparency setting isnt explicitly defined, default to on for all non batched things, gifs, and basis textures with alpha
-      switch (this.data.alphaMode) {
-        case "opaque":
-          this.mesh.material.transparent = false;
-          break;
-        case "blend":
-          this.mesh.material.transparent = true;
-          this.mesh.material.alphaTest = 0;
-          break;
-        case "mask":
-          this.mesh.material.transparent = false;
-          this.mesh.material.alphaTest = this.data.alphaCutoff;
-          break;
-        default:
-          this.mesh.material.transparent =
-            !this.data.batch ||
-            this.data.contentType.includes("image/gif") ||
-            !!(texture.image && texture.image.hasAlpha);
-          this.mesh.material.alphaTest = 0;
-      }
-    }
-
-    this.mesh.material.map = texture;
-    this.mesh.material.needsUpdate = true;
-
-    if (projection === "flat") {
-      scaleToAspectRatio(this.el, ratio);
-    }
-
-    if (texture !== errorTexture && this.data.batch && !texture.isCompressedTexture) {
-      batchManagerSystem.addObject(this.mesh);
-    }
-
-    this.el.emit("image-loaded", { src: this.data.src, projection: projection });
   }
 });
