+++ src/hub.js
@@ -97,6 +97,7 @@ import "./components/stats-plus";
 import "./components/networked-avatar";
 import "./components/media-views";
 import "./components/media-pdf";
+import "./components/media-image";
 import "./components/avatar-volume-controls";
 import "./components/pinch-to-move";
 import "./components/pitch-yaw-rotator";
