+++ src/utils/gif-texture.js
@@ -1,0 +1,34 @@
+export class GIFTexture extends THREE.Texture {
+  constructor(frames, delays, disposals) {
+    super(document.createElement("canvas"));
+    this.image.width = frames[0].width;
+    this.image.height = frames[0].height;
+
+    this._ctx = this.image.getContext("2d");
+
+    this.generateMipmaps = false;
+    this.isVideoTexture = true;
+    this.minFilter = THREE.NearestFilter;
+
+    this.frames = frames;
+    this.delays = delays;
+    this.disposals = disposals;
+
+    this.frame = 0;
+    this.frameStartTime = Date.now();
+  }
+
+  update() {
+    if (!this.frames || !this.delays || !this.disposals) return;
+    const now = Date.now();
+    if (now - this.frameStartTime > this.delays[this.frame]) {
+      if (this.disposals[this.frame] === 2) {
+        this._ctx.clearRect(0, 0, this.image.width, this.image.width);
+      }
+      this.frame = (this.frame + 1) % this.frames.length;
+      this.frameStartTime = now;
+      this._ctx.drawImage(this.frames[this.frame], 0, 0, this.image.width, this.image.height);
+      this.needsUpdate = true;
+    }
+  }
+}
